fix initial state on web
