+++ packages/expo-router/src/link/linking.ts
@@ -1,5 +1,6 @@
 import Constants, { ExecutionEnvironment } from "expo-constants";
 import * as Linking from "expo-linking";
+import { Platform } from "react-native";
 
 import getPathFromState from "../fork/getPathFromState";
 import getStateFromPath from "../fork/getStateFromPath";
@@ -10,7 +11,14 @@ import getStateFromPath from "../fork/getStateFromPath";
 // For example, if you had a root navigator where the first screen was `/settings` and the second was `/index`
 // then `/index` would be used on web and `/settings` would be used on native.
 export async function getInitialURL(): Promise<string> {
-  const url = await Promise.race<string>([
+  if (Platform.OS === "web") {
+    if (typeof window === "undefined") {
+      return "";
+    } else if (typeof window.location?.href === "string") {
+      return window.location.href;
+    }
+  }
+  return Promise.race<string>([
     (async () => {
       const url = await Linking.getInitialURL();
 
@@ -39,7 +47,6 @@ export async function getInitialURL(): Promise<string> {
       setTimeout(() => resolve(getRootURL()), 150)
     ),
   ]);
-  return url;
 }
 
 let _rootURL: string | undefined;
