+++ packages/expo-router/src/rootStateContext.tsx
@@ -1,42 +1,7 @@
-import React, { createContext, useContext, useEffect, useState } from "react";
-import URL from "url-parse";
+import React, { createContext, useContext } from "react";
 
 import { State } from "./fork/getPathFromState";
-import { useLinkingContext } from "./link/useLinkingContext";
-import { useServerState } from "./useServerState";
-
-function useResolvedPromise<T>(promise: Promise<T> | undefined) {
-  const [resolved, setResolved] = useState<T | undefined>();
-
-  useEffect(() => {
-    if (promise) {
-      promise.then(setResolved);
-    }
-  }, [promise]);
-
-  return resolved;
-}
-
-function useHackInitialRootState() {
-  // TODO: We probably don't need this
-  const serverState = useServerState();
-
-  const linking = useLinkingContext();
-  const url = useResolvedPromise(linking.getInitialURL() as Promise<string>);
-  const [state, setState] = useState<State | null>(null);
-
-  useEffect(() => {
-    if (url) {
-      const parsed = URL(url);
-      // TODO: Add hashes to the path
-      const urlWithoutOrigin = parsed.pathname + parsed.query;
-
-      setState(linking.getStateFromPath(urlWithoutOrigin, linking.config)!);
-    }
-  }, [url, linking]);
-
-  return state ?? serverState;
-}
+import { useInitialRootState } from "./useInitialRootState";
 
 export const InitialRootStateContext = createContext<State | null>(null);
 
@@ -59,11 +24,12 @@ export function InitialRootStateProvider({
 }: {
   children: React.ReactNode;
 }) {
-  const state = useHackInitialRootState();
+  const state = useInitialRootState();
 
-  // Prevent all rendering until we have the initial root state.
-  // Probably React Navigation should be doing this for us.
+  // TODO: This is only used on native but it shouldn't be used at all.
   if (!state) {
+    // Prevent all rendering until we have the initial root state.
+    // Probably React Navigation should be doing this for us.
     return null;
   }
