+++ packages/expo-router/src/useInitialRootState.tsx
@@ -1,0 +1,16 @@
+import { useLinkingContext } from "./link/useLinkingContext";
+import { useServerState } from "./useServerState";
+
+export function useInitialRootState() {
+  // Node.js environments
+  if (typeof window === "undefined") {
+    // eslint-disable-next-line react-hooks/rules-of-hooks
+    return useServerState();
+  }
+
+  // Web environments
+
+  // eslint-disable-next-line react-hooks/rules-of-hooks
+  const linking = useLinkingContext();
+  return linking.getStateFromPath(window.location.pathname, linking.config);
+}
